VirtualServer class creation implemented
VirtualServer instances are created by looping over the "server" blocks
in a configuration - contained in a Layer data structure.

It requires a pointer to the correct Layer, and requires the use of a
named constructor - which will first verify that the supplied Layer is
valid for use in a VirtualServer (It must contains a "server" block,
with a "listen" directive that has a valid port/socket_address).

A VirtualServer::address() function is also implemented, that will
create a SocketAddrV4 instance based on the address of the virtual
server.

It should be possible, once the VirtualServers and TcpListeners are
correctly set up, to:
1. Assign a SocketAddrV4 to a Client based on the TcpListener which
   accepted its connection
2. After a Request is completed, and the client has signalled that it is
   ready for a response, then before beginning the generation of a
   Response, the server will check which VirtualServer corresponds with
   the SocketAddrV4 of the client, and therefore use the appropriate
   configuration.