VirtualServer set-up complete (I think)
The VirtualServer class has been created - it contains a pointer to a
relevant Server block.

The negative consequence of using a pointer instead of the
Layer::Location struct I created before, is that now the configuration
Layer MUST be heap allocated, to ensure that the pointer is still valid.

Additionally, unlike Rust's references, the compiler is not going to help us out by
checking that the pointer is still valid. This shouldn't be a problem
however, as the configuration and the VirtualServer's should both have
the same lifetime, and indeed, both exist within the same scope - inside
of a WebServ::Config class.

For the first time in this commit, I've begun working with the main.cpp
of the program.

It is now set up to:
- Read a config file - either default.conf, or a file specified by
  argv[1] on the command line.
- Verify that the config file is valid - exit with an error if invalid.
- Create VirtualServer instances, and TcpListeners, based on the
  "server" blocks that exist within the configuration file. Additional
  error checking will take place here, and exit the program with an
  error if there is an issue. These instances will exist inside of a
  struct WebServ::Config.
- Once the WebServ::Config has been successfully constructed, it will be
  passed by reference to the WebServ::run() function, where the code for
  receiving Requests and processing them into appropriate responses will
  be placed.
- If an error should occur in WebServ::run() (which is shouldn't as per
  the project guidelines), then the error will be printed before the
  program exits.

Currently, a placeholder function for handling of clients will simply
print out whether a client's request was valid or not. If invalid, it
will give the reason. If valid, it will print out the well-formed
request.

There have not been sufficient Integration tests written for this code
yet, and this may be the next area to focus on.
All Unit tests are currently passing however, and valgrind shows no
memory issues or leaks.