Progress on ServerLogic
Testing to see if WebServ can actually be displayed in a browser window,
the result was good, but only for a hardcoded HTML string so far.

One important note was that without the Content-Length header, the
HTTP Response was ill formed, and caused issues on both Firefox and
Chrome. The Response Builder has now been altered to always add a
Content-Length header. In future, this may be omitted in the case of
chunked transfer encoding.