A lot of messy code
Code is anything but clean in this commit.

Changes:

- After a Request is completed, and the client requests information
back from the Server, then before the generation of a Response, the
correct 'server' block from the Configuration file is selected based on
the 'Host' header (or a host in the URI if present).

- Extra validations in several areas are present, including checking
to see whether every 'server' block has access to a 'root' directive (in
it's local scope or in the scope above),
which is required for proper functionality.

- Additional getters for Request and Layer classes added.

- URI getters modified to match prototype of Request getters.